# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatsapp Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://official.whatsapp.loc/engine"
        }
    ],
    "paths": {
...
        "/instances/qr": {
            "get": {
                "tags": [
                    "Instances"
                ],
                "summary": "fetch qr",
                "description": "fetch user a new qr or return connected as status if connection established",
                "operationId": "qr",
                "responses": {
                    "200": {
                        "description": "returns qr image base64 or connected."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/instances/status": {
            "get": {
                "tags": [
                    "Instances"
                ],
                "summary": "fetch qr",
                "description": "fetch user connection status",
                "operationId": "status",
                "responses": {
                    "200": {
                        "description": "returns disconnected or connected."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/instances/disconnect": {
            "post": {
                "tags": [
                    "Instances"
                ],
                "summary": "disconnect connection",
                "description": "disconnect user current connection",
                "operationId": "disconnect",
                "responses": {
                    "200": {
                        "description": "returns Disconnected."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/instances/clearInstance": {
            "post": {
                "tags": [
                    "Instances"
                ],
                "summary": "clear instance data",
                "description": "clear instance data",
                "operationId": "clearInstance",
                "responses": {
                    "200": {
                        "description": "returns Disconnected."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/instances/updateChannelSetting": {
            "post": {
                "tags": [
                    "Instances"
                ],
                "summary": "update instance setting",
                "description": "update instance settings like webhook urls",
                "operationId": "updateChannelSetting",
                "parameters": [
                    {
                        "name": "webhooks[messageNotifications]",
                        "in": "query",
                        "description": "Messages Webhook",
                        "required": false
                    },
                    {
                        "name": "webhooks[ackNotifications]",
                        "in": "query",
                        "description": "Messages status Ack Webhook",
                        "required": false
                    },
                    {
                        "name": "webhooks[chatNotifications]",
                        "in": "query",
                        "description": "Dialogs Webhook",
                        "required": false
                    },
                    {
                        "name": "webhooks[labelNotifications]",
                        "in": "query",
                        "description": "Labels Webhook",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns channel setting object after modifications."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/profiles/me": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "summary": "fetch user data",
                "description": "fetch connected user data",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "returns me object."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/profiles/updateProfilePicture": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update Profile Picture",
                "description": "update my profile picture or my groups",
                "operationId": "updateProfilePicture",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone or group id",
                        "required": true
                    },
                    {
                        "name": "imageURL",
                        "in": "query",
                        "description": "Image Url",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns updateProfilePicture object."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/profiles/updatePresence": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update Presence for 10 seconds",
                "description": "update target presence",
                "operationId": "updatePresence",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone or group id",
                        "required": true
                    },
                    {
                        "name": "presence",
                        "in": "query",
                        "description": "User Presence",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns updatePresence object."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/profiles/updateName": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update Profile Name",
                "description": "update profile name",
                "operationId": "updateName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns updateName object."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/profiles/updateStatus": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "summary": "Update Profile Status",
                "description": "update profile status",
                "operationId": "updateStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns updateStatus object."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/queues/getMessagesQueue": {
            "get": {
                "tags": [
                    "Queues"
                ],
                "summary": "fetch user messages queue",
                "description": "fetch user messages queue",
                "operationId": "getMessagesQueue",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/queues/clearMessagesQueue": {
            "post": {
                "tags": [
                    "Queues"
                ],
                "summary": "clear user messages queue",
                "description": "clear user messages queue",
                "operationId": "clearMessagesQueue",
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "channel_id": {
                "type": "apiKey",
                "description": "Enter Your Channel ID",
                "name": "CHANNELID",
                "in": "header"
            },
            "channel_token": {
                "type": "apiKey",
                "description": "Enter Your Channel Token",
                "name": "CHANNELTOKEN",
                "in": "header"
            }
        }
    }
}